REFACTOR: Separate DHCPS and DHCP
Move Message and MessageOption into the DHCP namespace
